+++ crates/config/src/toolchain/python_config.rs
@@ -1,8 +1,20 @@
-// use super::bin_config::BinEntry;
-use schematic::Config;
+use schematic::{derive_enum, Config, ConfigEnum};
 use serde::Serialize;
 use version_spec::UnresolvedVersionSpec;
 use warpgate_api::PluginLocator;
+
+#[cfg(feature = "proto")]
+use crate::inherit_tool;
+
+derive_enum!(
+    /// The available package managers for Python.
+    #[derive(ConfigEnum, Copy, Default)]
+    pub enum PythonPackageManager {
+        #[default]
+        Pip,
+        Uv,
+    }
+);
 
 #[derive(Clone, Config, Debug, PartialEq, Serialize)]
 pub struct PipConfig {
@@ -10,18 +22,38 @@ pub struct PipConfig {
     pub install_args: Option<Vec<String>>,
 }
 
+#[derive(Clone, Config, Debug, PartialEq, Serialize)]
+pub struct UvConfig {
+    /// List of arguments to append to `uv install` commands.
+    pub install_args: Option<Vec<String>>,
+
+    /// Location of the WASM plugin to use for uv support.
+    pub plugin: Option<PluginLocator>,
+
+    /// The version of uv to download, install, and run `uv` tasks with.
+    #[setting(env = "MOON_UV_VERSION")]
+    pub version: Option<UnresolvedVersionSpec>,
+}
+
 #[derive(Clone, Config, Debug, PartialEq)]
 pub struct PythonConfig {
-    /// Location of the WASM plugin to use for Python support.
-    pub plugin: Option<PluginLocator>,
+    /// The package manager to use for installing dependencies.
+    pub package_manager: PythonPackageManager,
 
     /// Options for pip, when used as a package manager.
     #[setting(nested)]
-    pub pip: Option<PipConfig>,
+    pub pip: PipConfig,
+
+    /// Location of the WASM plugin to use for Python support.
+    pub plugin: Option<PluginLocator>,
 
     /// Assumes only the root `requirements.txt` is used for dependencies.
     /// Can be used to support the "one version policy" pattern.
     pub root_requirements_only: bool,
+
+    /// Options for uv, when used as a package manager.
+    #[setting(nested)]
+    pub uv: Option<UvConfig>,
 
     /// Defines the virtual environment name, which will be created in the workspace root.
     /// Project dependencies will be installed into this.
@@ -31,4 +63,26 @@ pub struct PythonConfig {
     /// The version of Python to download, install, and run `python` tasks with.
     #[setting(env = "MOON_PYTHON_VERSION")]
     pub version: Option<UnresolvedVersionSpec>,
+}
+
+#[cfg(feature = "proto")]
+impl PythonConfig {
+    inherit_tool!(UvConfig, uv, "uv", inherit_proto_uv);
+
+    pub fn inherit_proto(&mut self, proto_config: &proto_core::ProtoConfig) -> miette::Result<()> {
+        match &self.package_manager {
+            PythonPackageManager::Pip => {
+                // Built-in
+            }
+            PythonPackageManager::Uv => {
+                if self.uv.is_none() {
+                    self.uv = Some(UvConfig::default());
+                }
+
+                self.inherit_proto_uv(proto_config)?;
+            }
+        }
+
+        Ok(())
+    }
 }